pyproject.toml
[tool.poetry]
name = "agean-cyber"
version = "0.1.0"
description = "An earnest and calculated attempt at building an active cyber agent with an RL reward"
authors = ["nentropy <nen@nentropic.dev>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.1"
langgraph = "^0.2.31"

python-dotenv = "^1.0.1"
langchain-community = "^0.3.1"
pyyaml = "^6.0.2"
sqlalchemy = "^2.0.35"
psycopg2 = "^2.9.9"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
elasticsearch = "^8.15.1"
logstash = "^0.1.dev0"
uuid = "^1.30"
datetime = "^5.5"
qdrant-client = "^1.11.3"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.18.2"
ollama = "^0.3.3"
google-cloud-secret-manager = "^2.20.2"
humanloop = "^0.8.2"
pyinputplus = "^0.2.12"
openai = "^1.51.0"
anthropic = "^0.34.2"
alembic = "^1.13.3"
fastapi = "^0.115.0"
langsmith = "^0.1.130"
pandas = "^2.2.3"
langchain-experimental = "^0.3.2"
matplotlib = "^3.9.2"
langchain-core = "^0.3.8"
chromadb = "^0.5.11"
asyncio = "^3.4.3"
llama-index = "^0.11.15"
portkey-ai = "^1.8.7"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pylint = "^3.3.1"
pip-tools = "^7.4.1"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
